Added and fixes something

+++ Content/NPCs/Bosses/CosmicJellyfish.cs
@@ -8,9 +8,12 @@ using ITD.Content.Items.Weapons.Summoner;
 using ITD.Content.Projectiles.Hostile;
 using ITD.Players;
 using ITD.Utilities;
+using Microsoft.Build.Evaluation;
+using Microsoft.CodeAnalysis;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using ReLogic.Content;
+using SteelSeries.GameSense;
 using System;
 using Terraria;
 using Terraria.Audio;
@@ -72,7 +75,9 @@ namespace ITD.Content.NPCs.Bosses
             Waiting,
             Charging,
             Slinging,
-            DownToSize
+            DownToSize,
+            VoidShard,
+            StarShard
         }
 
         private HandState handState = HandState.Waiting;
@@ -109,7 +114,19 @@ namespace ITD.Content.NPCs.Bosses
             notExpertRule.OnSuccess(ItemDropRule.Common(ModContent.ItemType<StarlitOre>(), 1, 15, 30));
             notExpertRule.OnSuccess(ItemDropRule.OneFromOptionsNotScalingWithLuck(1, oneFromOptionsDrops));
         }
+        public override void ApplyDifficultyAndPlayerScaling(int numPlayers, float balance, float bossAdjustment)
+        {
+            if (!Main.masterMode)
+            {
+                NPC.lifeMax = (int)(NPC.lifeMax * 0.8f * bossAdjustment);
+            }
+            else
+            {
+                NPC.lifeMax = (int)(NPC.lifeMax * 0.8f * bossAdjustment);
 
+            }
+            NPC.damage = (int)(NPC.damage * 0.7f);
+        }
         public override void OnKill()
         {
             DownedBossSystem.downedCosJel = true;
@@ -221,9 +238,17 @@ namespace ITD.Content.NPCs.Bosses
 
                         NPC.localAI[2] = 0;
                         NPC.ai[3]++;
+                        if (hand != -1)
+                        {
+                            handState = HandState.DownToSize;
+                        }
+                        if (hand2 != -1)
+                        {
+                            handState2 = HandState.DownToSize;
+                        }
                         //Gore MUST goes here (else it will fuck up multiplayer)
                     }
-                    Main.NewText("ACKK STUDIO II: CRUSTACEANDO. IN THE JELLY OF THE FISH.", Color.Crimson);
+                    Main.NewText("Fish Your Eyes", Color.Goldenrod);
                     NPC.netUpdate = true;
 
                 }
@@ -235,12 +260,9 @@ namespace ITD.Content.NPCs.Bosses
         private void Attacks(Player player)//___________________________________________________________________________________________________________________________________________________
         {
             //TODO: Change this??
-            float distanceAbove = 275f;
             Vector2 toPlayer = player.Center - NPC.Center;
             Vector2 toPlayerNormalized = Vector2.Normalize(toPlayer);
-            Vector2 abovePlayer = toPlayer + new Vector2(0f, -distanceAbove);
-            Vector2 aboveNormalized = Vector2.Normalize(abovePlayer);
-            float speed = abovePlayer.Length();
+
 
             if (NPC.ai[3] == 0)//No attack begin trasition
             {
@@ -254,14 +276,14 @@ namespace ITD.Content.NPCs.Bosses
             if (NPC.ai[3] == 1)//Sludge
             {
                 NPC.localAI[1]++;
-                if (NPC.localAI[1] == 180 || NPC.localAI[1] == 120 && SecondStage)
+                if (NPC.localAI[1] == 180 || NPC.localAI[1] == 60 && SecondStage)
                 {
                     if (Main.expertMode || Main.masterMode)
                     {
                         NPC.localAI[0]++;
                         if (NPC.localAI[0] < 2 && !SecondStage || NPC.localAI[0] < 3 && !SecondStage)
                         {
-                            NPC.localAI[1] = 90;
+                            NPC.localAI[1] = 20;
                         }
                         else
                         {
@@ -296,38 +318,17 @@ namespace ITD.Content.NPCs.Bosses
             }
             if (NPC.ai[3] == 2)//Crystal Meth
             {
-                NPC.localAI[1]++;
-                if (NPC.localAI[1] == 150 || (NPC.localAI[1] == 100 || NPC.localAI[1] == 150) && SecondStage)
+                if (NPC.localAI[1]++ >= 100)
                 {
-
-                    if (NPC.localAI[1]++ >= 150)
+                    NPC.localAI[1] = 0;
+                    if (handState == HandState.Waiting && hand != -1)
+                        handState = HandState.VoidShard;
+                }
+                else
+                {
+                    if (hand == -1)
                     {
-                        NPC.ai[3]++;
-                        NPC.localAI[1] = 0;
-                    }
-
-                    SoundEngine.PlaySound(SoundID.Item20, NPC.Center);
-                    //P2 stat garbage here
-                    int projectileAmount = Main.rand.Next(5, 11);
-                    float radius = 6.5f;
-                    float sector = MathHelper.ToRadians(80f);
-                    //Full circle
-                    if (Main.expertMode || Main.masterMode)
-                    {
-                        projectileAmount = Main.rand.Next(18, 24);
-                        sector =(float)(MathHelper.TwoPi);
-                    }
-                    float sectorOfSector = sector / projectileAmount;
-                    float towardsAngle = toPlayer.ToRotation();
-                    float startAngle = towardsAngle - sectorOfSector * (projectileAmount - 1) / 2;
-                    if (Main.netMode != NetmodeID.MultiplayerClient)
-                    {
-                        for (int i = 0; i < projectileAmount; i++)
-                        {
-                            float angle = startAngle + sectorOfSector * i;
-                            Vector2 projectileVelo = new Vector2((float)Math.Cos(angle), (float)Math.Sin(angle)) * radius;
-                            Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, projectileVelo, ModContent.ProjectileType<CosmicVoidShard>(), 20, 5, -1);
-                        }
+                        hand = Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, Vector2.Zero, ModContent.ProjectileType<CosmicJellyfish_Hand>(), 1, 0.1f);
                     }
                 }
             }
@@ -360,21 +361,19 @@ namespace ITD.Content.NPCs.Bosses
             }
             if (NPC.ai[3] == 4)//FTHOF
             {
-                if (NPC.localAI[1]++ >= 200)
-                {
-                    NPC.ai[3]++;
-                    NPC.localAI[1] = 0;
-                    if (handState2 == HandState.Waiting && hand2 != -1)
-                        handState2 = HandState.Charging;
-                    if (handState == HandState.Waiting && hand != -1)
-                        handState = HandState.Charging;
-                }
-                else
-                {
                     if (Main.netMode != NetmodeID.MultiplayerClient)
                     {
                         if (!SecondStage)
                         {
+                            if (NPC.localAI[1]++ >= 200)
+                            {
+                                NPC.ai[3]++;
+                                NPC.localAI[1] = 0;
+                                if (handState2 == HandState.Waiting && hand2 != -1)
+                                    handState2 = HandState.Charging;
+                                if (handState == HandState.Waiting && hand != -1)
+                                    handState = HandState.Charging;
+                            }
                             if (hand == -1 && hand2 == -1)
                             {
                                 if (Main.rand.NextBool(2))
@@ -395,6 +394,21 @@ namespace ITD.Content.NPCs.Bosses
                         }
                         else
                         {
+                        if (NPC.localAI[1]++ == 100)
+                        {
+                            if (handState == HandState.Waiting && hand != -1)
+                                handState = HandState.Charging;
+                        }
+                        if (NPC.localAI[2] >6)
+                        {
+                            NPC.localAI[1] = 0;
+                            NPC.localAI[2] = 0;
+                            NPC.ai[3]++;
+                            if (handState == HandState.Waiting && hand != -1)
+                                handState = HandState.DownToSize;
+                            if (handState2 == HandState.Waiting && hand2 != -1)
+                                handState2 = HandState.DownToSize;
+                        }
                             if (hand == -1)
                             {
                                 hand = Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, Vector2.Zero, ModContent.ProjectileType<CosmicJellyfish_Hand>(), 1, 0.1f);
@@ -406,8 +420,6 @@ namespace ITD.Content.NPCs.Bosses
                         }
 
                     }
-
-                }
             }
                 if (NPC.ai[3] == 5)//Suffocate, actual retarded attack but i'm not the boss here
             {
@@ -457,11 +469,14 @@ namespace ITD.Content.NPCs.Bosses
                          ModContent.ProjectileType<CosmicRayWarn>(), NPC.damage, 0f, -1, 240, NPC.whoAmI);
                     }
                 }
-                else if (NPC.localAI[1] >= 300)
+                else if (NPC.localAI[1] >= 180)
                 {
                     NPC.ai[3]++;
                     NPC.localAI[1] = 0;
+                    NPC.localAI[2] = 0;
+                    NPC.localAI[0] = 0;
                     NPC.netUpdate = true;
+
                 }
             }
             if (NPC.ai[3] == -1)//Subterranean Sun
@@ -553,6 +568,22 @@ namespace ITD.Content.NPCs.Bosses
                     Vector2 chargedPosition = NPC.Center - extendOut * 110 + new Vector2(0f, NPC.velocity.Y) - toPlayer * 150f;
                     Vector2 normalCenter = NPC.Center - extendOut * 110 + new Vector2(0f, NPC.velocity.Y);
                     float targetRotation = NPC.rotation;
+                    if (handState != HandState.Charging && handState != HandState.Slinging)
+                    {
+                        if (projectile.frameCounter++ >= 6)
+                        {
+                            projectile.frameCounter = 0;
+                            projectile.frame++;
+                            if (projectile.frame >= 4)
+                            {
+                                projectile.frame =0;
+                            }
+                        }
+                    }
+                    else
+                    {
+                            projectile.frame = 5;
+                    }
                     switch (handState)
                     {
                         case HandState.Waiting:
@@ -574,7 +605,7 @@ namespace ITD.Content.NPCs.Bosses
                                 {
                                     LockedIn = player.Center;
                                 }
-                                else if (Timer >= 5 && SecondStage || Timer >= 0 && !SecondStage)
+                                else if (Timer >= 5 && NPC.localAI[2] > 6 || Timer >= 0 && NPC.localAI[2] <=6)
                                 {
                                     Timer = 0;
                                     handState = HandState.Slinging;
@@ -601,7 +632,15 @@ namespace ITD.Content.NPCs.Bosses
                                 }
                                 else
                                 {
-                                    projectile.Kill();
+                                    if (NPC.localAI[2]++ <= 6 && SecondStage)
+                                    {
+                                        if (handState2 == HandState.Waiting && hand2 != -1)
+                                            handState2 = HandState.Charging;
+                                    }
+                                    if (NPC.localAI[2] > 6 || !SecondStage)
+                                    {
+                                        projectile.Kill();
+                                    }
                                     handState = HandState.Waiting;
                                 }
                             }
@@ -617,6 +656,45 @@ namespace ITD.Content.NPCs.Bosses
                                 handState = HandState.Waiting;
                                 projectile.Kill();
                             }
+                            break;
+                        case HandState.VoidShard:
+                            if (NPC.localAI[0]++>= 100)
+                            {
+                                NPC.localAI[0] = 0;
+                                if (NPC.localAI[2]++ >= 5)
+                                {
+                                    NPC.ai[3]++;
+                                    handState = HandState.DownToSize;
+                                    NPC.localAI[2] = 0;
+                                    NPC.localAI[1] = 0;
+                                    NPC.localAI[0] = 0;
+                                }
+
+                                SoundEngine.PlaySound(SoundID.Item20, projectile.Center);
+                                //P2 stat garbage here
+                                int projectileAmount = Main.rand.Next(5, 11);
+                                float radius = 6.5f;
+                                float sector = MathHelper.ToRadians(80f);
+                                //Full circle
+                                if (Main.expertMode || Main.masterMode)
+                                {
+                                    projectileAmount = Main.rand.Next(16, 20);
+                                    sector = (float)(MathHelper.TwoPi);
+                                }
+                                float sectorOfSector = sector / projectileAmount;
+                                float towardsAngle = toPlayer.ToRotation();
+                                float startAngle = towardsAngle - sectorOfSector * (projectileAmount - 1) / 2;
+                                if (Main.netMode != NetmodeID.MultiplayerClient)
+                                {
+                                    for (int i = 0; i < projectileAmount; i++)
+                                    {
+                                        float angle = startAngle + sectorOfSector * i;
+                                        Vector2 projectileVelo = new Vector2((float)Math.Cos(angle), (float)Math.Sin(angle)) * radius;
+                                        Projectile.NewProjectile(projectile.GetSource_FromAI(), projectile.Center, projectileVelo, ModContent.ProjectileType<CosmicVoidShard>(), 20, 5, -1);
+                                    }
+                                }
+                            }
+                            projectile.Center = Vector2.Lerp(projectile.Center, normalCenter, 0.3f);
                             break;
                     }
                     projectile.rotation = MathHelper.Lerp(projectile.rotation, targetRotation, 0.3f);
@@ -666,6 +744,23 @@ namespace ITD.Content.NPCs.Bosses
                     Vector2 chargedPosition = NPC.Center - extendOut * -110 + new Vector2(0f, NPC.velocity.Y) - toPlayer * 150f;
                     Vector2 normalCenter = NPC.Center - extendOut * -110 + new Vector2(0f, NPC.velocity.Y);
                     float targetRotation = NPC.rotation;
+                    if (handState2 != HandState.Charging && handState2 != HandState.Slinging)
+                    {
+                        if (projectile2.frameCounter++ >= 6)
+                        {
+                            projectile2.frameCounter = 0;
+                            projectile2.frame++;
+                            if (projectile2.frame >= 4)
+                            {
+
+                                projectile2.frame = 0;
+                            }
+                        }
+                    }
+                    else
+                    {
+                            projectile2.frame = 5;
+                    }
                     switch (handState2)
                     {
                         case HandState.Waiting:
@@ -685,7 +780,7 @@ namespace ITD.Content.NPCs.Bosses
                             {
                                 if (Timer2++ == 0)
                                 {
-                                    LockedIn = player.Center;
+                                        LockedIn = player.Center;
                                 }
                                 else if (Timer2 >= 5 && SecondStage || Timer2 >= 0 && !SecondStage)
                                 {
@@ -714,7 +809,15 @@ namespace ITD.Content.NPCs.Bosses
                                 }
                                 else
                                 {
-                                    projectile2.Kill();
+                                    if (NPC.localAI[2]++ <= 6 && SecondStage)
+                                    {
+                                        if (handState == HandState.Waiting && hand != -1)
+                                            handState = HandState.Charging;
+                                    }
+                                    if (NPC.localAI[2] > 6 || !SecondStage)
+                                    {
+                                        projectile2.Kill();
+                                    }
                                     handState2 = HandState.Waiting;
                                 }
                                 
@@ -771,13 +874,13 @@ namespace ITD.Content.NPCs.Bosses
         //a fuck you when despawn is epic
         private void Movement(Player player)//___________________________________________________________________________________________________________________________________________________
         {
-            float distanceAbove = 275f;
+            float distanceAbove = 250f;//True melee
             Vector2 toPlayer = player.Center - NPC.Center;
             Vector2 toPlayerNormalized = Vector2.Normalize(toPlayer);
             Vector2 abovePlayer = toPlayer + new Vector2(0f, -distanceAbove);
             Vector2 aboveNormalized = Vector2.Normalize(abovePlayer);
             Vector2 dashvel;
-            float speed = abovePlayer.Length();
+            float speed = abovePlayer.Length()/1.3f;//True melee
             switch (AI_State)
             {
                 case MovementState.FollowingRegular:
@@ -874,6 +977,14 @@ namespace ITD.Content.NPCs.Bosses
                 NPC.dontTakeDamage = true;
                 NPC.netUpdate = true;
                 DesperateAttack = true;
+                if (hand != -1)
+                {
+                    handState = HandState.DownToSize;
+                }
+                if (hand2 != -1)
+                {
+                    handState2 = HandState.DownToSize;
+                }
                 AI_State = MovementState.Explode;
                 Main.NewText("Subterranean Sun.", Color.Orange);
                 return false;
